added: WatchedPolicy, tests, stubs

+++ app/Http/Controllers/Watched/WatchedController.php
@@ -6,29 +6,34 @@ use App\Http\Controllers\Controller;
 use App\Models\Movies\Movie;
 use App\Models\Watched\Watched;
 use Illuminate\Http\Request;
+use Illuminate\Support\Arr;
 
 class WatchedController extends Controller
 {
-    protected $base_view_path = 'movies.watched';
+    protected $base_view_path = 'watchable.watched';
 
     /**
      * Display a listing of the resource.
      *
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @return \Illuminate\Http\Response
      */
-    public function index(Movie $movie)
+    public function index(Request $request, string $type, Movie $watchable)
     {
-        //
+        if ($request->wantsJson()) {
+            return $watchable->watched()
+                ->latest()
+                ->paginate();
+        }
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @return \Illuminate\Http\Response
      */
-    public function create(Movie $movie)
+    public function create(string $type, Movie $watchable)
     {
         //
     }
@@ -37,15 +42,16 @@ class WatchedController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request, Movie $movie)
+    public function store(Request $request, string $type, Movie $watchable)
     {
-        $watched = $movie->watched()->create([
-            'user_id' => auth()->id(),
-            'watched_at' => now(),
+        $attributes = $request->validate([
+            'watched_at' => 'sometimes|date',
         ]);
+
+        $watched = $watchable->watchedBy(auth()->user(), $attributes);
 
         return $watched;
     }
@@ -53,12 +59,14 @@ class WatchedController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @param  \App\Models\Watched\Watched  $watched
      * @return \Illuminate\Http\Response
      */
-    public function show(Request $request, Movie $movie, Watched $watched)
+    public function show(Request $request, string $type, Movie $watchable, Watched $watched)
     {
+        $this->authorize('view', $watched);
+
         if ($request->wantsJson()) {
             return $watched;
         }
@@ -67,25 +75,27 @@ class WatchedController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @param  \App\Models\Watched\Watched  $watched
      * @return \Illuminate\Http\Response
      */
-    public function edit(Movie $movie, Watched $watched)
+    public function edit(string $type, Movie $watchable, Watched $watched)
     {
-        //
+        $this->authorize('update', $watched);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @param  \App\Models\Watched\Watched  $watched
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Movie $movie, Watched $watched)
+    public function update(Request $request, string $type, Movie $watchable, Watched $watched)
     {
+        $this->authorize('update', $watched);
+
         $attributes = $request->validate([
             'watched_at' => 'required|date',
         ]);
@@ -108,12 +118,14 @@ class WatchedController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Movies\Movie  $movie
+     * @param  \App\Models\Movies\Movie  $watchable
      * @param  \App\Models\Watched\Watched  $watched
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request, Movie $movie, Watched $watched)
+    public function destroy(Request $request, string $type, Movie $watchable, Watched $watched)
     {
+        $this->authorize('delete', $watched);
+
         if ($is_deletable = $watched->isDeletable()) {
             $watched->delete();
         }
@@ -137,7 +149,7 @@ class WatchedController extends Controller
             ];
         }
 
-        return redirect(route($this->base_view_path . '.index', ['movie' => $movie->id]))
+        return redirect(route($this->base_view_path . '.index', ['type' => 'movies', 'watchable' => $watchable->id]))
             ->with('status', $status);
     }
 }
